# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/students": {
      "get": {
        "description": "Returns a paginated list of students visible to the caller. Scope rules per role:\n- **principal / vice_principal**: all active students in the school.\n- **supervisor**: students in halaqat the supervisor oversees.\n- **teacher**: students in halaqat where the teacher is primary.\n- **parent**: their own children across all schools (the one exception to strict school-scoping).\n\n`include_deleted=true` is silently ignored for non-principal/VP callers. Does NOT write an audit log (read-only).",
        "operationId": "StudentsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Defaults to 1.",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Defaults to 20, capped at 100.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Name search",
            "schema": {
              "example": "يوسف",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status. Omit to return all statuses.",
            "schema": {
              "enum": [
                "active",
                "inactive",
                "graduated"
              ],
              "type": "string"
            }
          },
          {
            "name": "gender",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "male",
                "female"
              ],
              "type": "string"
            }
          },
          {
            "name": "halaqa_id",
            "required": false,
            "in": "query",
            "description": "Filter by halaqa. For supervisors/teachers this is an additional scope filter on top of their role-based visibility.",
            "schema": {
              "example": 3,
              "type": "number"
            }
          },
          {
            "name": "include_deleted",
            "required": false,
            "in": "query",
            "description": "Principal/VP only",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "id_number",
            "required": false,
            "in": "query",
            "description": "Exact-match filter on national ID number (normalized before comparison).",
            "schema": {
              "maxLength": 20,
              "example": "300123456",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentListEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List students (role-scoped)",
        "tags": [
          "Students"
        ]
      },
      "post": {
        "description": "Creates a student in the caller's school. An optional `guardians[]` array links guardians atomically in the same transaction — the first entry is forced to `is_primary=true` regardless of the submitted value. For the email-branch guardian flow, if no user exists a new account is created with the `parent` role and a parent-invite email is dispatched. Audit: `student.create` (+ `student.guardian.link` per linked guardian).",
        "operationId": "StudentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient role",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a student (BR-STU-01)",
        "tags": [
          "Students"
        ]
      }
    },
...
    "/students/{id}/error-heatmap": {
      "get": {
        "description": "Aggregates the student's recitation errors by ayah over the last `days` days, returning the worst ayat (most errors first) with per-type counts. Staff only — parents do not see error breakdowns. Out-of-scope or cross-school access returns 404.",
        "operationId": "StudentErrorHeatmapController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Student ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "description": "How many days back to aggregate errors over. Defaults to 56 (8 weeks).",
            "schema": {
              "minimum": 1,
              "maximum": 365,
              "default": 56,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max number of hotspots (worst ayat) to return. Defaults to 20.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorHeatmapEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Student not found or out of scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a student's error heatmap",
        "tags": [
          "Students"
        ]
      }
    },
...
  },
  "info": {
    "title": "Halaqa Backend API",
    "description": "Authentication, user management, and school operations for the Halaqa system. All JSON responses are wrapped in a uniform envelope: `{ \"code\": <http>, \"data\": ... }` on success, `{ \"code\": <http>, \"message\": \"...\" }` on errors. Refresh tokens travel as HttpOnly cookies on `/auth/*`; access tokens use Bearer auth.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 401
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          },
          "details": {
            "description": "Validation error list (only on 400 with multiple field errors).",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "StudentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "يوسف محمد"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ],
            "example": "male"
          },
          "dob": {
            "type": "object",
            "nullable": true,
            "format": "date",
            "example": "2015-06-01"
          },
          "join_date": {
            "type": "string",
            "format": "date",
            "example": "2024-09-01"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "graduated"
            ],
            "example": "active"
          },
          "daily_hifz_pages_capacity": {
            "type": "string",
            "example": "1.00",
            "description": "Daily Hifz memorisation target in pages (0–20)."
          },
          "daily_near_pages_capacity": {
            "type": "string",
            "example": "2.00",
            "description": "Daily near-review target in pages (0–50)."
          },
          "daily_far_pages_capacity": {
            "type": "string",
            "example": "5.00",
            "description": "Daily far-review target in pages (0–100)."
          },
          "notes": {
            "type": "object",
            "nullable": true,
            "example": "Needs extra support on Juz 30."
          },
          "photo_url": {
            "type": "object",
            "nullable": true,
            "example": "https://cdn.example.com/photos/1.jpg"
          },
          "id_number": {
            "type": "object",
            "nullable": true,
            "example": "300123456",
            "description": "National ID number. Always present when you can see the student. Value is null for legacy students created before this field was required."
          }
        },
        "required": [
          "id",
          "name",
          "gender",
          "dob",
          "join_date",
          "status",
          "daily_hifz_pages_capacity",
          "daily_near_pages_capacity",
          "daily_far_pages_capacity",
          "notes",
          "photo_url",
          "id_number"
        ]
      },
      "StudentListData": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentResponse"
            }
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
      "StudentListEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/StudentListData"
          }
        },
        "required": [
          "code",
          "data"
        ]
      },
...
      "LinkGuardianDto": {
        "type": "object",
        "properties": {
          "guardian_user_id": {
            "type": "number",
            "example": 42,
            "description": "ID of an existing user. Mutually exclusive with `email`. Must belong to the same school."
          },
          "email": {
            "type": "string",
            "example": "parent@example.com",
            "format": "email",
            "description": "Email of an existing user. Mutually exclusive with `guardian_user_id`. Returns 404 if not found."
          },
          "relation": {
            "type": "string",
            "enum": [
              "father",
              "mother",
              "grandfather",
              "grandmother",
              "uncle",
              "aunt",
              "sibling",
              "other"
            ],
            "example": "father"
          },
          "is_primary": {
            "type": "boolean",
            "example": true,
            "description": "Exactly one guardian per student may be primary. If omitted on the first guardian, forced to true. Setting true unsets the current primary. Setting false on the current primary is rejected (400)."
          },
          "can_pickup": {
            "type": "boolean",
            "example": true,
            "description": "Whether the guardian is authorised to pick up the student. Defaults to true."
          }
        },
        "required": [
          "relation"
        ]
      },
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "يوسف محمد",
            "minLength": 2,
            "maxLength": 100
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female"
            ],
            "example": "male"
          },
          "dob": {
            "type": "string",
            "format": "date",
            "example": "2015-06-01",
            "nullable": true
          },
          "join_date": {
            "type": "string",
            "format": "date",
            "example": "2024-09-01"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "graduated"
            ],
            "example": "active",
            "description": "Defaults to `active` when omitted."
          },
          "daily_hifz_pages_capacity": {
            "type": "number",
            "example": 1,
            "minimum": 0,
            "maximum": 20,
            "description": "Daily Hifz memorisation target in pages. Defaults to 0.5."
          },
          "daily_near_pages_capacity": {
            "type": "number",
            "example": 2,
            "minimum": 0,
            "maximum": 50,
            "description": "Daily near-review target in pages. Defaults to 2."
          },
          "daily_far_pages_capacity": {
            "type": "number",
            "example": 5,
            "minimum": 0,
            "maximum": 100,
            "description": "Daily far-review target in pages. Defaults to 5."
          },
          "notes": {
            "type": "string",
            "example": "Needs extra support on Juz 30."
          },
          "photo_url": {
            "type": "string",
            "example": "https://cdn.example.com/photos/1.jpg"
          },
          "id_number": {
            "type": "string",
            "example": "300123456",
            "maxLength": 20,
            "description": "National ID number. Format is country-specific (Palestinian: 9 digits). A bad checksum is stored with a warning, not rejected. Unique per school (soft-deleted rows included in the constraint)."
          },
          "guardians": {
            "description": "Guardians linked atomically in the same transaction. First entry is forced to is_primary=true.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkGuardianDto"
            }
          }
        },
        "required": [
          "name",
          "gender",
          "join_date",
          "id_number"
        ]
      },
      "StudentEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/StudentResponse"
          },
          "warnings": {
            "example": [
              "id_number.checksum_invalid"
            ],
            "description": "Non-fatal warnings. Present only when id_number validation produced warnings.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "code",
          "data"
        ]
      },
...
      "HeatmapHotspotDto": {
        "type": "object",
        "properties": {
          "surah": {
            "type": "number",
            "example": 2
          },
          "ayah": {
            "type": "number",
            "example": 5
          },
          "mistakes_count": {
            "type": "number",
            "example": 3
          },
          "warnings_count": {
            "type": "number",
            "example": 1
          },
          "tajweed_errors_count": {
            "type": "number",
            "example": 0
          },
          "harakat_errors_count": {
            "type": "number",
            "example": 2
          },
          "total": {
            "type": "number",
            "example": 6,
            "description": "Total errors at this ayah in the window."
          }
        },
        "required": [
          "surah",
          "ayah",
          "mistakes_count",
          "warnings_count",
          "tajweed_errors_count",
          "harakat_errors_count",
          "total"
        ]
      },
      "ErrorHeatmapResponse": {
        "type": "object",
        "properties": {
          "student_id": {
            "type": "number",
            "example": 42
          },
          "days": {
            "type": "number",
            "example": 56,
            "description": "The window (in days) that was aggregated."
          },
          "total_errors": {
            "type": "number",
            "example": 37,
            "description": "Total error occurrences across all ayat in the window."
          },
          "hotspots": {
            "description": "Worst ayat, most errors first (ties broken by mushaf order).",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HeatmapHotspotDto"
            }
          }
        },
        "required": [
          "student_id",
          "days",
          "total_errors",
          "hotspots"
        ]
      },
      "ErrorHeatmapEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/ErrorHeatmapResponse"
          }
        },
        "required": [
          "code",
          "data"
        ]
      },
...
    }
  }
}